replace spree_multi_site with 2.0 stable

+++ spree_multi_site/db/migrate/20121216093739_last_fix_unique_index.rb
@@ -1,15 +1,16 @@
 class LastFixUniqueIndex < ActiveRecord::Migration
   def change
     if Spree::Site.table_exists?
-      
-      remove_index "spree_users", :name => "email_idx_unique"
-      add_index "spree_users", ["site","email"], :name => "email_idx_unique", :unique => true
-      
-      
+      if defined?(Spree::Auth)
+        remove_index "spree_users", :name => "email_idx_unique"      
+        add_index "spree_users", ["site_id","email"], :name => "email_idx_unique", :unique => true
+      end
+            
       remove_index :spree_preferences, :name => 'index_spree_preferences_on_key'
-      add_index "spree_preferences", ["site","key"], :name => "index_spree_preferences_on_key", :unique => true
+      add_index "spree_preferences", ["site_id","key"], :name => "index_spree_preferences_on_key", :unique => true
 
+      remove_index :spree_products, :name => 'permalink_idx_unique'
+      add_index "spree_products", ["site_id", "permalink"], :name => "permalink_idx_unique", :unique => true
     end
   end
-
 end
